Reduce the amount of packages
