+++ SSLTest/src/run.py
@@ -5,7 +5,7 @@ import sys
 import traceback
 
 from .scan import handle_scan_output
-from .scan_parameters.non_ratable.port_discovery import discover_ports
+from .parameters.port_discovery import discover_ports
 
 log = logging.getLogger(__name__)
