+++ SSLTest/src/vulnerabilities/VulnerabilityTest.py
@@ -2,7 +2,7 @@ import concurrent.futures as cf
 import logging
 from abc import ABC, abstractmethod
 
-from ..scan_parameters.connections.connection_utils import worst_or_best_protocol
+from ..parameters.connection_utils import worst_or_best_protocol
 from ..utils import protocol_version_conversion, Address
 
 log = logging.getLogger(__name__)
