+++ SSLTest/src/vulnerabilities/tests/HSTSSupport.py
